+++ ktorfit-ksp/src/main/kotlin/de/jensklingenberg/ktorfit/reqBuilderExtension/MethodCodeGeneration.kt
@@ -4,14 +4,10 @@ import de.jensklingenberg.ktorfit.model.annotations.CustomHttp
 import de.jensklingenberg.ktorfit.model.annotations.HttpMethodAnnotation
 
 fun getMethodCode(httpMethod: HttpMethodAnnotation): String {
-    val httpMethodValue = when (httpMethod) {
-        is CustomHttp -> {
-            httpMethod.customValue
-        }
-
-        else -> {
-            httpMethod.httpMethod.keyword
-        }
+    val httpMethodValue = if (httpMethod is CustomHttp) {
+        httpMethod.customValue
+    } else {
+        httpMethod.httpMethod.keyword
     }
     return "method = HttpMethod.parse(\"${httpMethodValue}\")"
 }