+++ ktorfit-lib-common/src/commonMain/kotlin/de/jensklingenberg/ktorfit/converter/builtin/KtorfitDefaultConverterFactory.kt
@@ -3,7 +3,7 @@ package de.jensklingenberg.ktorfit.converter.builtin
 import de.jensklingenberg.ktorfit.Ktorfit
 import de.jensklingenberg.ktorfit.Response
 import de.jensklingenberg.ktorfit.converter.Converter
-import de.jensklingenberg.ktorfit.converter.KtorfitResponse
+import de.jensklingenberg.ktorfit.converter.KtorfitResult
 import de.jensklingenberg.ktorfit.internal.TypeData
 import io.ktor.client.call.*
 import io.ktor.client.statement.*
@@ -43,36 +43,14 @@ internal class KtorfitDefaultConverterFactory : Converter.Factory {
             }
         }
 
-        override suspend fun convert(ktorfitResponse: KtorfitResponse): Response<Any?> {
-            val typeInfo = typeData.typeArgs.first().typeInfo
-            return when (ktorfitResponse) {
-                is KtorfitResponse.Success -> {
-                    val rawResponse = ktorfitResponse.response
-
-                    val code: Int = rawResponse.status.value
-                    when {
-                        code < 200 || code >= 300 -> {
-                            val errorBody = rawResponse.body<Any>()
-                            Response.error(errorBody, rawResponse)
-                        }
-
-                        code == 204 || code == 205 -> {
-                            Response.success<Any>(null, rawResponse)
-                        }
-
-                        else -> {
-                            val convertedBody = ktorfit.nextSuspendResponseConverter(
-                                null,
-                                typeData.typeArgs.first()
-                            )?.convert(ktorfitResponse)
-                                ?: rawResponse.body<Any>(typeInfo)
-                            Response.success(convertedBody, rawResponse)
-                        }
-                    }
+        override suspend fun convert(ktorfitResult: KtorfitResult): Response<Any?> {
+            return when (ktorfitResult) {
+                is KtorfitResult.Success -> {
+                    convert(ktorfitResult.response)
                 }
 
-                is KtorfitResponse.Failed -> {
-                    Response.error(null,null,ktorfitResponse.throwable)
+                is KtorfitResult.Failed -> {
+                    throw ktorfitResult.throwable
                 }
             }
         }