+++ sandbox/src/commonMain/kotlin/com/example/model/ExampleApi.kt
@@ -1,6 +1,5 @@
 package com.example.model
 
-import de.jensklingenberg.ktorfit.Call
 import de.jensklingenberg.ktorfit.Response
 import de.jensklingenberg.ktorfit.http.GET
 
@@ -9,8 +8,5 @@ interface ExampleApi {
     suspend fun getUser(): Response<User>
 
     @GET("example.json")
-    suspend fun getCallUser(): Call<User>
-
-    @GET("example.json")
-    suspend fun getCallUse3r()
+    suspend fun getUserMy(): MyOwnResponse<User>
 }