+++ sandbox/src/commonMain/kotlin/com/example/model/MyOwnResponse.kt
@@ -2,12 +2,13 @@ package com.example.model
 
 import kotlinx.serialization.Serializable
 
-@Serializable sealed class MyOwnResponse<T> {
+@Serializable
+sealed class MyOwnResponse<T> {
     data class Success<T>(val data: T) : MyOwnResponse<T>()
-    class Error(val ex:Throwable) : MyOwnResponse<Nothing>()
+    class Error<T>(val ex: Throwable) : MyOwnResponse<Nothing>()
 
     companion object {
         fun <T> success(data: T) = Success(data)
-        fun error(ex: Throwable) = Error(ex)
+        fun error(ex: Throwable) = Error<Nothing>(ex)
     }
 }