+++ sandbox/src/commonMain/kotlin/com/example/model/MyOwnResponseConverter.kt
@@ -2,31 +2,11 @@ package com.example.model
 
 import de.jensklingenberg.ktorfit.Ktorfit
 import de.jensklingenberg.ktorfit.converter.Converter
-import de.jensklingenberg.ktorfit.converter.SuspendResponseConverter
+import de.jensklingenberg.ktorfit.converter.KtorfitResult
 import de.jensklingenberg.ktorfit.internal.TypeData
 import io.ktor.client.call.*
 import io.ktor.client.statement.*
-import io.ktor.util.reflect.*
 
-class MyOwnResponseConverter : SuspendResponseConverter {
-
-    override suspend fun <RequestType> wrapSuspendResponse(
-        typeData: TypeData,
-        requestFunction: suspend () -> Pair<TypeInfo, HttpResponse>,
-        ktorfit: Ktorfit
-    ): Any {
-        return try {
-            val (info, response) = requestFunction()
-            MyOwnResponse.success<Any>(response.body(info))
-        } catch (ex: Throwable) {
-            MyOwnResponse.error(ex)
-        }
-    }
-
-    override fun supportedType(typeData: TypeData, isSuspend: Boolean): Boolean {
-        return typeData.qualifiedName == "com.example.model.MyOwnResponse"
-    }
-}
 
 class MyOwnResponseConverterFactory : Converter.Factory {
 
@@ -35,13 +15,30 @@ class MyOwnResponseConverterFactory : Converter.Factory {
         ktorfit: Ktorfit
     ): Converter.SuspendResponseConverter<HttpResponse, *>? {
         if (typeData.typeInfo.type == MyOwnResponse::class) {
+
             return object : Converter.SuspendResponseConverter<HttpResponse, Any> {
                 override suspend fun convert(response: HttpResponse): Any {
                     return try {
-                        val data = response.call.body(typeData.typeArgs.first().typeInfo)
-                        MyOwnResponse.success(data)
+                        val convertedBody = ktorfit.nextSuspendResponseConverter(
+                            null,
+                            typeData.typeArgs.first()
+                        )?.convert(response)
+                            ?: response.body(typeData.typeArgs.first().typeInfo)
+                        MyOwnResponse.success(convertedBody)
                     } catch (ex: Throwable) {
                         MyOwnResponse.error(ex)
+                    }
+                }
+
+                override suspend fun convert(ktorfitResult: KtorfitResult): Any {
+                    return when (ktorfitResult) {
+                        is KtorfitResult.Failed -> {
+                            MyOwnResponse.error(ktorfitResult.throwable)
+                        }
+
+                        is KtorfitResult.Success -> {
+                            convert(ktorfitResult.response)
+                        }
                     }
                 }
             }