+++ sandbox/src/jvmMain/kotlin/de/jensklingenberg/ktorfit/demo/JvMMain.kt
@@ -4,8 +4,8 @@ package de.jensklingenberg.ktorfit.demo
 import com.example.UserFactory
 import com.example.api.JsonPlaceHolderApi
 import com.example.model.ExampleApi
-import com.example.model.User
-import de.jensklingenberg.ktorfit.Callback
+import com.example.model.MyOwnResponse
+import com.example.model.MyOwnResponseConverterFactory
 import de.jensklingenberg.ktorfit.converter.builtin.CallConverterFactory
 import de.jensklingenberg.ktorfit.converter.builtin.FlowConverterFactory
 import de.jensklingenberg.ktorfit.ktorfit
@@ -47,7 +47,11 @@ val userKtorfit = ktorfit {
     baseUrl("https://foso.github.io/Ktorfit/")
     httpClient(jvmClient)
 
-    converterFactories(FlowConverterFactory(), UserFactory(), CallConverterFactory())
+    converterFactories(
+        FlowConverterFactory(),
+        MyOwnResponseConverterFactory(),
+        UserFactory(), CallConverterFactory()
+    )
 }
 
 
@@ -55,13 +59,16 @@ fun main() {
 
     runBlocking {
 
-        val user = userKtorfit.create<ExampleApi>().getUser()
+        val user = userKtorfit.create<ExampleApi>().getUserMy()
 
-        if (user.isSuccessful) {
-            println(user.body())
-        } else {
-            user.errorBody()
-            println(user.failure().toString())
+        when (user) {
+            is MyOwnResponse.Success -> {
+                println(user.data)
+            }
+
+            is MyOwnResponse.Error<*> -> {
+                println(user.ex)
+            }
         }
         delay(3000)
     }